Rename world/router directory to world/routing to improve comprehension.
